Updates and bugfixes for OrderBook

+++ src/model/market/books/OrderBook.java
@@ -21,7 +21,7 @@ public interface OrderBook {
 	// Returns total transaction value for 'quantity' units, if successful.
 	// Otherwise, throws LiquidityException, which contains the number
 	// successfully executed.
-	public double executeMarketOrder(OrderType type, double quantity) throws LiquidityException;
+	public double executeMarketOrder(OrderType type, int quantity) throws LiquidityException;
 
 	// Place a limit order for a particular price/quantity/expirationTime
 	// returns true if order was successfully placed (-not- executed) and false
@@ -50,10 +50,10 @@ public interface OrderBook {
 
 	public double[] getSellOrders();
 
-	// The current best price offered by buyers
+	// The current best (lowest) price offered by sellers
 	public double getAskPrice();
 
-	// The current best price offered by sellers
+	// The current best (highest) price offered by buyers
 	public double getBidPrice();
 
 	// The gap between the best price
